+++ precompiles/ics20/tx.go
@@ -14,8 +14,8 @@ import (
 
 	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/evmos/evmos/v20/x/evm/config"
 	"github.com/evmos/evmos/v20/x/evm/core/vm"
+	evmtypes "github.com/evmos/evmos/v20/x/evm/types"
 
 	cmn "github.com/evmos/evmos/v20/precompiles/common"
 )
@@ -69,7 +69,7 @@ func (p *Precompile) Transfer(
 		return nil, err
 	}
 
-	if contract.CallerAddress != origin && msg.Token.Denom == config.GetEVMCoinDenom() {
+	if contract.CallerAddress != origin && msg.Token.Denom == evmtypes.GetEVMCoinDenom() {
 		// escrow address is also changed on this tx, and it is not a module account
 		// so we need to account for this on the UpdateDirties
 		escrowAccAddress := transfertypes.GetEscrowAddress(msg.SourcePort, msg.SourceChannel)