+++ rpc/backend/node_info.go
@@ -26,7 +26,7 @@ import (
 	rpctypes "github.com/evmos/evmos/v20/rpc/types"
 	"github.com/evmos/evmos/v20/server/config"
 	"github.com/evmos/evmos/v20/types"
-	evmconfig "github.com/evmos/evmos/v20/x/evm/config"
+	evmtypes "github.com/evmos/evmos/v20/x/evm/types"
 )
 
 // Accounts returns the list of accounts available to this node.
@@ -281,7 +281,7 @@ func (b *Backend) GenerateMinGasCoin(gasPrice hexutil.Big, appConf config.Config
 
 	// fetch the base denom from the sdk Config in case it's not currently defined on the node config
 	if len(minGasPrices) == 0 || minGasPrices.Empty() {
-		unit = evmconfig.GetEVMCoinDenom()
+		unit = evmtypes.GetEVMCoinDenom()
 	} else {
 		unit = minGasPrices[0].Denom
 	}
@@ -335,7 +335,7 @@ func (b *Backend) RPCBlockRangeCap() int32 {
 // the node config. If set value is 0, it will default to 20.
 
 func (b *Backend) RPCMinGasPrice() int64 {
-	baseDenom := evmconfig.GetEVMCoinDenom()
+	baseDenom := evmtypes.GetEVMCoinDenom()
 
 	minGasPrice := b.cfg.GetMinGasPrices()
 	amt := minGasPrice.AmountOf(baseDenom).TruncateInt64()