+++ x/evm/wrappers/utils_test.go
@@ -11,7 +11,7 @@ import (
 	"cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/evmos/evmos/v20/types"
-	"github.com/evmos/evmos/v20/x/evm/config"
+	evmtypes "github.com/evmos/evmos/v20/x/evm/types"
 	"github.com/evmos/evmos/v20/x/evm/wrappers"
 
 	"github.com/stretchr/testify/require"
@@ -22,42 +22,42 @@ func TestMustConvertEvmCoinTo18Decimals(t *testing.T) {
 
 	testCases := []struct {
 		name        string
-		evmCoinInfo config.EvmCoinInfo
+		evmCoinInfo evmtypes.EvmCoinInfo
 		coin        sdk.Coin
 		expCoin     sdk.Coin
 		expPanic    bool
 	}{
 		{
 			name:        "pass - zero amount 18 decimals",
-			evmCoinInfo: config.EvmCoinInfo{Denom: types.BaseDenom, Decimals: config.EighteenDecimals},
+			evmCoinInfo: evmtypes.EvmCoinInfo{Denom: types.BaseDenom, Decimals: evmtypes.EighteenDecimals},
 			coin:        baseCoinZero,
 			expPanic:    false,
 			expCoin:     baseCoinZero,
 		},
 		{
 			name:        "pass - zero amount 6 decimals",
-			evmCoinInfo: config.EvmCoinInfo{Denom: types.BaseDenom, Decimals: config.SixDecimals},
+			evmCoinInfo: evmtypes.EvmCoinInfo{Denom: types.BaseDenom, Decimals: evmtypes.SixDecimals},
 			coin:        baseCoinZero,
 			expPanic:    false,
 			expCoin:     baseCoinZero,
 		},
 		{
 			name:        "pass - no conversion with 18 decimals",
-			evmCoinInfo: config.EvmCoinInfo{Denom: types.BaseDenom, Decimals: config.EighteenDecimals},
+			evmCoinInfo: evmtypes.EvmCoinInfo{Denom: types.BaseDenom, Decimals: evmtypes.EighteenDecimals},
 			coin:        sdk.Coin{Denom: types.BaseDenom, Amount: math.NewInt(10)},
 			expPanic:    false,
 			expCoin:     sdk.Coin{Denom: types.BaseDenom, Amount: math.NewInt(10)},
 		},
 		{
 			name:        "pass - conversion with 6 decimals",
-			evmCoinInfo: config.EvmCoinInfo{Denom: types.BaseDenom, Decimals: config.SixDecimals},
+			evmCoinInfo: evmtypes.EvmCoinInfo{Denom: types.BaseDenom, Decimals: evmtypes.SixDecimals},
 			coin:        sdk.Coin{Denom: types.BaseDenom, Amount: math.NewInt(1)},
 			expPanic:    false,
 			expCoin:     sdk.Coin{Denom: types.BaseDenom, Amount: math.NewInt(1e12)},
 		},
 		{
 			name:        "panic - not evm denom",
-			evmCoinInfo: config.EvmCoinInfo{Denom: types.BaseDenom, Decimals: config.SixDecimals},
+			evmCoinInfo: evmtypes.EvmCoinInfo{Denom: types.BaseDenom, Decimals: evmtypes.SixDecimals},
 			coin:        sdk.Coin{Denom: "evmos", Amount: math.NewInt(1)},
 			expPanic:    true,
 		},
@@ -77,7 +77,9 @@ func TestMustConvertEvmCoinTo18Decimals(t *testing.T) {
 				}
 			}()
 
-			config.SetEVMCoinTEST(tc.evmCoinInfo)
+			configurator := evmtypes.NewEVMConfigurator()
+			configurator.ResetTestChainConfig()
+			require.NoError(t, configurator.WithEVMCoinInfo(tc.evmCoinInfo.Denom, tc.evmCoinInfo.Decimals).Configure())
 
 			coinConverted := wrappers.MustConvertEvmCoinTo18Decimals(tc.coin)
 
@@ -93,47 +95,47 @@ func TestConvertEvmCoinFrom18Decimals(t *testing.T) {
 
 	testCases := []struct {
 		name        string
-		evmCoinInfo config.EvmCoinInfo
+		evmCoinInfo evmtypes.EvmCoinInfo
 		coin        sdk.Coin
 		expCoin     sdk.Coin
 		expErr      bool
 	}{
 		{
 			name:        "pass - zero amount 18 decimals",
-			evmCoinInfo: config.EvmCoinInfo{Denom: types.BaseDenom, Decimals: config.EighteenDecimals},
+			evmCoinInfo: evmtypes.EvmCoinInfo{Denom: types.BaseDenom, Decimals: evmtypes.EighteenDecimals},
 			coin:        baseCoinZero,
 			expErr:      false,
 			expCoin:     baseCoinZero,
 		},
 		{
 			name:        "pass - zero amount 6 decimals",
-			evmCoinInfo: config.EvmCoinInfo{Denom: types.BaseDenom, Decimals: config.SixDecimals},
+			evmCoinInfo: evmtypes.EvmCoinInfo{Denom: types.BaseDenom, Decimals: evmtypes.SixDecimals},
 			coin:        baseCoinZero,
 			expErr:      false,
 			expCoin:     baseCoinZero,
 		},
 		{
 			name:        "pass - no conversion with 18 decimals",
-			evmCoinInfo: config.EvmCoinInfo{Denom: types.BaseDenom, Decimals: config.EighteenDecimals}, coin: sdk.Coin{Denom: types.BaseDenom, Amount: math.NewInt(10)}, expErr: false,
+			evmCoinInfo: evmtypes.EvmCoinInfo{Denom: types.BaseDenom, Decimals: evmtypes.EighteenDecimals}, coin: sdk.Coin{Denom: types.BaseDenom, Amount: math.NewInt(10)}, expErr: false,
 			expCoin: sdk.Coin{Denom: types.BaseDenom, Amount: math.NewInt(10)},
 		},
 		{
 			name:        "pass - conversion with 6 decimals",
-			evmCoinInfo: config.EvmCoinInfo{Denom: types.BaseDenom, Decimals: config.SixDecimals},
+			evmCoinInfo: evmtypes.EvmCoinInfo{Denom: types.BaseDenom, Decimals: evmtypes.SixDecimals},
 			coin:        sdk.Coin{Denom: types.BaseDenom, Amount: math.NewInt(1e12)},
 			expErr:      false,
 			expCoin:     sdk.Coin{Denom: types.BaseDenom, Amount: math.NewInt(1)},
 		},
 		{
 			name:        "pass - conversion with amount less than conversion factor",
-			evmCoinInfo: config.EvmCoinInfo{Denom: types.BaseDenom, Decimals: config.SixDecimals},
+			evmCoinInfo: evmtypes.EvmCoinInfo{Denom: types.BaseDenom, Decimals: evmtypes.SixDecimals},
 			coin:        sdk.Coin{Denom: types.BaseDenom, Amount: math.NewInt(1e11)},
 			expErr:      false,
 			expCoin:     baseCoinZero,
 		},
 		{
 			name:        "fail - not evm denom",
-			evmCoinInfo: config.EvmCoinInfo{Denom: types.BaseDenom, Decimals: config.SixDecimals},
+			evmCoinInfo: evmtypes.EvmCoinInfo{Denom: types.BaseDenom, Decimals: evmtypes.SixDecimals},
 			coin:        sdk.Coin{Denom: "evmos", Amount: math.NewInt(1)},
 			expErr:      true,
 		},
@@ -141,7 +143,9 @@ func TestConvertEvmCoinFrom18Decimals(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			config.SetEVMCoinTEST(tc.evmCoinInfo)
+			configurator := evmtypes.NewEVMConfigurator()
+			configurator.ResetTestChainConfig()
+			require.NoError(t, configurator.WithEVMCoinInfo(tc.evmCoinInfo.Denom, tc.evmCoinInfo.Decimals).Configure())
 
 			coinConverted, err := wrappers.ConvertEvmCoinFrom18Decimals(tc.coin)
 
@@ -161,37 +165,37 @@ func TestConvertCoinsFrom18Decimals(t *testing.T) {
 
 	testCases := []struct {
 		name        string
-		evmCoinInfo config.EvmCoinInfo
+		evmCoinInfo evmtypes.EvmCoinInfo
 		coins       sdk.Coins
 		expCoins    sdk.Coins
 	}{
 		{
 			name:        "pass - no evm denom",
-			evmCoinInfo: config.EvmCoinInfo{Denom: types.BaseDenom, Decimals: config.SixDecimals},
+			evmCoinInfo: evmtypes.EvmCoinInfo{Denom: types.BaseDenom, Decimals: evmtypes.SixDecimals},
 			coins:       sdk.Coins{nonBaseCoin},
 			expCoins:    sdk.Coins{nonBaseCoin},
 		},
 		{
 			name:        "pass - only base denom 18 decimals",
-			evmCoinInfo: config.EvmCoinInfo{Denom: types.BaseDenom, Decimals: config.EighteenDecimals},
+			evmCoinInfo: evmtypes.EvmCoinInfo{Denom: types.BaseDenom, Decimals: evmtypes.EighteenDecimals},
 			coins:       sdk.Coins{baseCoin},
 			expCoins:    sdk.Coins{baseCoin},
 		},
 		{
 			name:        "pass - only base denom 6 decimals",
-			evmCoinInfo: config.EvmCoinInfo{Denom: types.BaseDenom, Decimals: config.SixDecimals},
+			evmCoinInfo: evmtypes.EvmCoinInfo{Denom: types.BaseDenom, Decimals: evmtypes.SixDecimals},
 			coins:       sdk.Coins{baseCoin},
 			expCoins:    sdk.Coins{sdk.Coin{Denom: types.BaseDenom, Amount: baseCoin.Amount.QuoRaw(1e12)}},
 		},
 		{
 			name:        "pass - multiple coins and base denom 18 decimals",
-			evmCoinInfo: config.EvmCoinInfo{Denom: types.BaseDenom, Decimals: config.EighteenDecimals},
+			evmCoinInfo: evmtypes.EvmCoinInfo{Denom: types.BaseDenom, Decimals: evmtypes.EighteenDecimals},
 			coins:       sdk.Coins{nonBaseCoin, baseCoin},
 			expCoins:    sdk.Coins{nonBaseCoin, baseCoin},
 		},
 		{
 			name:        "pass - multiple coins and base denom 6 decimals",
-			evmCoinInfo: config.EvmCoinInfo{Denom: types.BaseDenom, Decimals: config.SixDecimals},
+			evmCoinInfo: evmtypes.EvmCoinInfo{Denom: types.BaseDenom, Decimals: evmtypes.SixDecimals},
 			coins:       sdk.Coins{nonBaseCoin, baseCoin},
 			expCoins:    sdk.Coins{nonBaseCoin, sdk.Coin{Denom: types.BaseDenom, Amount: baseCoin.Amount.QuoRaw(1e12)}},
 		},
@@ -199,7 +203,9 @@ func TestConvertCoinsFrom18Decimals(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			config.SetEVMCoinTEST(tc.evmCoinInfo)
+			configurator := evmtypes.NewEVMConfigurator()
+			configurator.ResetTestChainConfig()
+			require.NoError(t, configurator.WithEVMCoinInfo(tc.evmCoinInfo.Denom, tc.evmCoinInfo.Decimals).Configure())
 
 			coinConverted := wrappers.ConvertCoinsFrom18Decimals(tc.coins)
 			require.Equal(t, tc.expCoins, coinConverted, "expected a different coin")
@@ -245,15 +251,17 @@ func TestZeroExtraDecimalsBigInt(t *testing.T) {
 		},
 	}
 
-	for _, cfg := range []config.EvmCoinInfo{
-		{Denom: types.BaseDenom, Decimals: config.SixDecimals},
-		{Denom: types.BaseDenom, Decimals: config.EighteenDecimals},
+	for _, cfg := range []evmtypes.EvmCoinInfo{
+		{Denom: types.BaseDenom, Decimals: evmtypes.SixDecimals},
+		{Denom: types.BaseDenom, Decimals: evmtypes.EighteenDecimals},
 	} {
 		for _, tc := range testCases {
 			t.Run(fmt.Sprintf("%d dec - %s", cfg.Decimals, tc.name), func(t *testing.T) {
-				config.SetEVMCoinTEST(cfg)
+				configurator := evmtypes.NewEVMConfigurator()
+				configurator.ResetTestChainConfig()
+				require.NoError(t, configurator.WithEVMCoinInfo(cfg.Denom, cfg.Decimals).Configure())
 				res := wrappers.AdjustExtraDecimalsBigInt(tc.amt)
-				if cfg.Decimals == config.EighteenDecimals {
+				if cfg.Decimals == evmtypes.EighteenDecimals {
 					tc.exp = tc.amt
 				}
 				require.Equal(t, tc.exp, res)